.x modified

+++ .x
@@ -16,6 +16,7 @@ EXTS=wiki
 TODO=TODO
 FAIL=FAIL
 WORK=git
+SKIP=SKIP
 
 # always work in the current directory
 # XXX TODO XXX add some config to more easily work in a working directory?
@@ -41,7 +42,7 @@ do
 		b="${b%".$c"}"
 	done
 	case "$b" in
-	(''|*[^a-zA-Z0-9.-]*)	printf 'NO: %q\n' "$a";;
+	(''|*[^a-zA-Z0-9.-_]*)	printf 'NO: %q\n' "$a";;
 	(*)			touch "$TODO/clone/$b";;
 	esac
 done
@@ -175,6 +176,7 @@ do
 	b="${a##*/}"
 
 	[ -d "$WORK/$b.git" ] || continue
+	[ -f "$SKIP/$b.git" ] && continue
 
 	try "$a" "$WORK/$b.git" pull "$b" &&
 	touch "$TODO/push/$b" &&
@@ -185,6 +187,7 @@ done
 for a in "$TODO/push/"*
 do
 	b="${a##*/}"
+	[ -f "$SKIP/$b.git" ] && continue
 
 	try "$a" "$WORK/$b.git" push "$b" &&
 	rm -f "$a"
@@ -217,7 +220,7 @@ then
 	done
 	printf '\n'
 	printf 'Hint: If you need to hack the remote name, do:\n'
-	printf 'while read repo dest; do ( cd %q/"$repo.git" && git config --local "url.%q${dest##*/}.insteadOf" %q"$repo".git ); done\n' "$WORK" "$DEST" "$DEST"
+	printf 'while read repo dest; do ( repo="${repo%%/}"; repo="${repo%%.git}"; repo="${repo%%%%*/}"; cd %q/"$repo.git" && git config --local "url.%q${dest##*/}.insteadOf" %q"$repo".git ); done\n' "$WORK" "$DEST" "$DEST"
 	printf 'To list those, do: make show\n'
 fi
